logger/requirements.txt
google-cloud-logging>=1.10.0
pyaml>=18.11.0
slacker-log-handler>=1.7.1